Move AnnotationPalette control functions into a separate file
This is so as to reduce the number of lines of code in CanvasViewModel.
